tests: move session joining tests to their own process
Fixes: #39
