+++ Cargo.toml
@@ -16,8 +16,6 @@ members = ["keyutils-raw"]
 [dev-dependencies]
 lazy_static = "1"
 regex = "1"
-serial_test = "*"
-serial_test_derive = "*"
 semver = "*"
 
 [dependencies]
